Split json generation from rails hash generation
When generating a hash with to_h, we want to leave native ruby types for
other ruby code to consume.

For json generation, add a new `as_json` method that returns a hash
designed to be passed to the `JSON.generate` method, which performs type
conversion - timestamp formatting in particular.